Add insertTemplate helper
